наконец вывел формулу NSWE относительно 2 координат, теперь можно играть с векторами

+++ src/main/java/org/example/CalcPosition/Point2D.java
@@ -1,10 +1,10 @@
 package org.example.CalcPosition;
 
 public class Point2D {
-    public int x;
-    public int y;
+    public double x;
+    public double y;
 
-    public Point2D(int x, int y) {
+    public Point2D(double x, double y) {
         this.x = x;
         this.y = y;
     }
